beyond scope of assignment
I think it's bad practice to rely on Pipfile.lock too much. It's simple
to use pip's show command to find out which version of packages were
installed and add this to the Pipfile as well. Perhaps this is a habit
from pip-tools? I'll need to research this further and add the content
to the Pipenv module if it's a valid concern.